+++ meerschaum/connectors/redis/__init__.py
@@ -1,0 +1,8 @@
+#! /usr/bin/env python3
+# vim:fenc=utf-8
+
+"""
+Import the `RedisConnector`.
+"""
+
+from meerschaum.connectors.redis.RedisConnector import RedisConnector